Don't keep appending listen.php to the URL

+++ src/logger/rx_extract.cxx
@@ -327,7 +327,7 @@ uint16_t gps_CRC16_checksum(char *s)
 	uint16_t x;
 	
 	for(x = 0xFFFF; *s; s++)
-		x = crc_xmodem_update(x, (uint16_t) *s);
+		x = crc_xmodem_update(x, (uint8_t) *s);
 	
 	return(x);
 }
